hakethon/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.7.0
psycopg[binary]
dj-database-url==2.2.0
requests==2.32.3
